# api-gateway/src/main/java/org/tinkoff/apigateway/controller/openapi-auth.yaml
openapi: "3.0.3"
info:
  title: "API-Gateway OpenAPI definition"
  description: "T-Eats API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
    description: "Generated server url"
paths:
  /auth/register:
    post:
      tags:
        - "auth-controller"
      operationId: "register"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /auth/login:
    post:
      tags:
        - "auth-controller"
      operationId: "login"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /auth/reset-password:
    post:
      tags:
        - "auth-controller"
      operationId: "resetPassword"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
  /auth/logout:
    post:
      tags:
        - "auth-controller"
      operationId: "logout"
      parameters:
        - name: "Authorization"
          in: "header"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "object"
components:
  schemas:
    RegisterRequest:
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
        email:
          type: "string"
        firstName:
          type: "string"
        lastName:
          type: "string"
    LoginRequest:
      type: "object"
      properties:
        username:
          type: "string"
        password:
          type: "string"
        rememberMe:
          type: "boolean"
    ResetPasswordRequest:
      type: "object"
      properties:
        username:
          type: "string"
        newPassword:
          type: "string"
        confirmationCode:
          type: "string"